Merges up -R into yarn up --recursive
Wires --recursive (and -R) as a boolean flag on the up command so
arguments can appear in any order. Previously --recursive was unsupported
and -R only worked as a path token, which forced it to come right after
the subcommand name.

Fixes 1 test.
